jqLiteExtras is a part of ui.scroll module now
